some minor renaming and preparation for the explorer
