--- StoryExplorer/ContextProvider.cs
@@ -1,6 +1,0 @@
-﻿namespace HousePartyTranslator
-{
-    internal class ContextProvider
-    {
-    }
-}
