+++ StoryExplorerForm/StoryExplorer.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Newtonsoft.Json;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
@@ -8,13 +9,16 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace HousePartyTranslator.StoryExplorer
+namespace HousePartyTranslator.StoryExplorerForm
 {
     public partial class StoryExplorer : Form
     {
+        private readonly ContextProvider Context;
         public StoryExplorer()
         {
             InitializeComponent();
+            Context = new ContextProvider("", ExplorerPanel);
+            Context.ParseFile();
         }
     }
 }
