+++ StoryExplorerForm/StoryExplorer.resx
